+++ HISTORY.md
@@ -2,14 +2,15 @@ Release History
 ===============
 1.3 (2021-02-28)
 -----------------
-- Support save artifacts as zip or only one file.
-- Support pythonic way to iterate jenkins and folder with depth.
+- Add capability to get/save artifacts for `WorkflowRun`.
+- Make `Jenkins` and `Folder` is subscribed and can be iterated with depth.
+- Refactor some code.
 
 1.2 (2021-01-31)
 ----------------
 - Support to enable, disable, scan, get_scan_log for `WorkflowMultiBranchProject`
 - Call `Jenkins.get_job` for getting parent of `Job`
-- Support process input step for WorkflowRun,  see `WorkflowRun.get_pending_input()`
+- Support process input step for `WorkflowRun`,  see `WorkflowRun.get_pending_input()`
 - Support user management
 
 1.1 (2020-12-31)